Fixes a race condition in reagent metabolism (#86509)
## About The Pull Request

Fixes a race condition that was causing organ damage to occur when it
shouldn't be.

The species `handle_chemical()` procs as well as the liver version of
that same proc get called in `metabolize_reagent()`, and in those procs
some toxins like carpotoxin get neutered.

Problem was, `metabolize_reagent()` was not being called until after the
organ damage calculation was already made using the un-neutered `toxpwr`
value.

Due to the way things are set up, the only option was to move the liver
damage code into `metabolize()`. The code itself is functionally
unchanged, it just has been relocated. This was a straightforward
process and actually eliminated a for loop (as well as some duplicated
code), so things may run a bit faster too. I did my best to keep it
legible as possible.

Fixes https://github.com/NovaSector/NovaSector/issues/2326 - this is
what got me to investigate, the issue doesn't come up as often on TG but
it becomes very obvious that something is not right when you have a
species' own exotic blood causing organ damage when it's just supposed
to be getting removed and added to blood_volume. This fixes that issue
and any other similar ones.

## Why It's Good For The Game

Species-specific damage immunities and such will now be respected.

## Changelog

:cl:
fix: fixed a race condition that was causing carpotoxin to cause liver
damage to felinids despite being immune
/:cl:

---------

Co-authored-by: SyncIt21 <110812394+SyncIt21@users.noreply.github.com>