Fish electrogenesis now scales with size. (#86593)
## About The Pull Request
A little touch to fish electrogenesis. Now its power scales with the
size of the fish. The initial values for the zip zap and the jumpercable
have been edited so that their average output is more or less
unaffected.

This PR also adds a check to fish petting to ensure that the user is
shocked only if the fish actually has the electrogenesis trait and not
just the energy that'd be generated if the fish had the trait.

## Why It's Good For The Game
The size and weight of a fish are variables that influence a lot of
mechanics such as reagents, number of fillets, weight class, force,
slowdown when carried or dragged, other traits, cargo value.
Electrogenesis should be no different.

Also fixing a possible oversight (not in the changelog because so far
the only fish with over 15 MJ is the zip zap and it's about impossible
to remove their electrogenesis trait rn also because they don't
reproduce).

## Changelog

:cl:
balance: Fish electrogenesis now scales with size.
/:cl: